Spent the day debugging issue with Consensus Append Entry. Resolved issue. Had to write sister project which uses NamedPipes to export sub millisecond accurate logs to trace bug.

+++ RaftConsensus/RaftConsensus/Common/Logging/RaftLogging.cs
@@ -1,49 +1,144 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.IO;
+using System.IO.Pipes;
 using System.Text;
 
 namespace TeamDecided.RaftConsensus.Common.Logging
 {
     public sealed class RaftLogging : IRaftLogging, IDisposable
     {
+        //Singleton
         public static RaftLogging Instance { get; } = new RaftLogging();
 
-        public event EventHandler<Tuple<ERaftLogType, string>> OnNewLogEntry;
+        //Member variables
+        private readonly object _methodLockObject = new object();
+        private ERaftLogType _logLevel;
+        public bool WriteToFile;
+        public bool WriteToNamedPipe;
+        public bool WriteToEvent;
+        private DateTime _startTime;
+        private Stopwatch _stopwatch;
 
-        private readonly  object _methodLockObject = new object();
-
-        private const string DefaultFilename = "debug.log";
-        private string _loggingFileName = DefaultFilename;
-
+        //File handling
         private List<string> _buffer;
         private int _linesToBufferCount;
+        public const string DefaultFilename = "debug.log";
+        private string _loggingFileName;
 
-        private ERaftLogType _logLevel;
+        //Event handling
+        public event EventHandler<Tuple<ERaftLogType, string>> OnNewLogEntry;
 
-        private RaftLogging() { }
+        //Named pipe handling
+        private NamedPipeClientStream _namedPipe;
+
+
+        private RaftLogging()
+        {
+            _startTime = DateTime.Now;
+            _stopwatch = new Stopwatch();
+            _stopwatch.Start();
+        }
 
         public void Log(ERaftLogType logType, string format, params object[] args)
         {
             lock (_methodLockObject)
             {
                 if (logType < _logLevel) return;
-
                 string message = string.Format(GetTimestampString() + format + Environment.NewLine, args);
 
-                OnNewLogEntry?.Invoke(this, new Tuple<ERaftLogType, string>(logType, message));
-
-                if (_buffer == null)
+                if (WriteToEvent)
                 {
-                    File.AppendAllText(_loggingFileName, message);
+                    WriteEvent(logType, message);
                 }
-                else
+
+                if (WriteToNamedPipe)
                 {
-                    _buffer.Add(message.TrimEnd());
-                    if (_buffer.Count == _linesToBufferCount)
-                    {
-                        FlushBuffer();
-                    }
+                    WriteNamedPipe(message);
+                }
+
+                if (WriteToFile)
+                {
+                    WriteFile(message);
+                }
+            }
+        }
+
+        private void WriteEvent(ERaftLogType logType, string message)
+        {
+            OnNewLogEntry?.Invoke(this, new Tuple<ERaftLogType, string>(logType, message));
+        }
+
+        private void WriteNamedPipe(string message)
+        {
+            SetupNamedPipe();
+            if(!WriteToNamedPipe)  return;
+
+            byte[] output = Encoding.UTF8.GetBytes(message);
+
+            try
+            {
+                _namedPipe.Write(output, 0, output.Length);
+                _namedPipe.Flush();
+            }
+            catch
+            {
+                SetupNamedPipe();
+                _namedPipe.Write(output, 0, output.Length);
+                _namedPipe.Flush();
+            }
+        }
+
+        public void NamedPipeRequestNewFile()
+        {
+            //Just a random sequence pattern for both sides to detect a direct message
+            WriteNamedPipe("###---###$$$MakeNewFile###---###$$$");
+        }
+
+        public void NamedPipeWriteToConsole(string message)
+        {
+            WriteNamedPipe("###---###$$$Message:" + message + "###---###$$$");
+        }
+
+        private void SetupNamedPipe()
+        {
+            if (_namedPipe == null)
+            {
+                _namedPipe = new NamedPipeClientStream(".", "RaftConsensusLogging", PipeDirection.Out,
+                    PipeOptions.WriteThrough);
+            }
+
+            if (_namedPipe.IsConnected) return;
+            try
+            {
+                _namedPipe.Connect(0);
+            }
+            catch (TimeoutException) //the server's named pipe is not running
+            {
+                WriteToNamedPipe = false;
+                _namedPipe.Dispose();
+                _namedPipe = null;
+            }
+        }
+
+        private void WriteFile(string message)
+        {
+            if (_loggingFileName == null)
+            {
+                return;
+            }
+
+            if (_buffer == null)
+            {
+                File.AppendAllText(_loggingFileName, message);
+            }
+            else
+            {
+                _buffer.Add(message.TrimEnd());
+                if (_buffer.Count == _linesToBufferCount)
+                {
+                    FlushBuffer();
                 }
             }
         }
@@ -88,6 +183,11 @@ namespace TeamDecided.RaftConsensus.Common.Logging
         {
             lock (_methodLockObject)
             {
+                if (_loggingFileName == null)
+                {
+                    return;
+                }
+
                 File.Delete(_loggingFileName);
             }
         }
@@ -105,15 +205,20 @@ namespace TeamDecided.RaftConsensus.Common.Logging
         {
             lock (_methodLockObject)
             {
+                if (_loggingFileName == null)
+                {
+                    return;
+                }
+
                 if (_buffer == null || _buffer.Count <= 0) return;
                 File.AppendAllLines(_loggingFileName, _buffer);
                 _buffer.Clear();
             }
         }
 
-        private static string GetTimestampString()
+        private string GetTimestampString()
         {
-            return DateTime.Now.ToString("HH:mm:ss.ffff") + ": ";
+            return (_startTime + _stopwatch.Elapsed).ToString("HH:mm:ss.fff") + ": ";
         }
 
         public static string FlattenException(Exception exception)
@@ -134,6 +239,9 @@ namespace TeamDecided.RaftConsensus.Common.Logging
         public void Dispose()
         {
             FlushBuffer();
+            _namedPipe?.Close();
+            _namedPipe?.Dispose();
+            _namedPipe = null;
         }
     }
 }
